+++ react/typings/vtex.product-context.d.ts
@@ -7,3 +7,13 @@ declare module 'vtex.product-context' {
 
   export const ProductContext = Context
 }
+
+declare module 'vtex.product-context/useProduct' {
+  import { Context } from 'react'
+
+  interface ProductContext {
+    product: any
+  }
+
+  export default function useProduct(): ProductContext
+}