+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Reviews Structured Data
+
+### Changed
+- Change code structure to use app settings from runtime, instead of fetching.
 
 ## [1.5.10] - 2020-07-14
 ### Fixed